Overall Changes

+++ app/src/main/java/de/fhe/ai/aurumbanking/view/deposit/DepositFragment.kt
@@ -3,7 +3,6 @@ package de.fhe.ai.aurumbanking.view.deposit
 import android.annotation.SuppressLint
 import android.os.Build
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -14,6 +13,7 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import de.fhe.ai.aurumbanking.R
 import de.fhe.ai.aurumbanking.core.Helper
+import java.math.BigDecimal
 
 class DepositFragment : Fragment(){
 
@@ -50,7 +50,7 @@ class DepositFragment : Fragment(){
     }
 
     @SuppressLint("SetTextI18n")
-    private fun setCustomerDepotToFragment(depositValue : Float) {
+    private fun setCustomerDepotToFragment(depositValue : BigDecimal) {
         this.root.findViewById<TextView?>(R.id.userDepotLeft).text = "${depositValue.toString()}\nEuro"
     }
 }
