+++ app/src/main/java/de/fhe/ai/aurumbanking/view/overview/OverviewViewModel.java
@@ -6,6 +6,8 @@ import androidx.annotation.NonNull;
 import androidx.lifecycle.AndroidViewModel;
 import androidx.lifecycle.LiveData;
 
+import java.math.BigDecimal;
+
 import de.fhe.ai.aurumbanking.storage.customer.CustomerRepository;
 import de.fhe.ai.aurumbanking.storage.deposit.DepositRepository;
 
@@ -27,7 +29,7 @@ public class OverviewViewModel extends AndroidViewModel {
 
     }
 
-    public LiveData<Float> getCustomerDepositByCustomerId(Long id){
+    public LiveData<BigDecimal> getCustomerDepositByCustomerId(Long id){
         return  depositRepository.getCustomerDepositByCustomerId(id);
     }
